chore: move objects to pkg
